requirements.txt
kiteconnect>=5.0.1
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.32.0
plotly>=5.18.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.36
ta>=0.11.0
openpyxl>=3.1.0
selenium>=4.18.0
webdriver-manager>=4.0.1
pyotp>=2.9.0
